Split render functions to utils to be used from other projects
