+++ src/sketch.ts
@@ -2,7 +2,8 @@ import p5 from "p5"
 import { Id, Interpreter } from "wollok-ts"
 import { DEFAULT_GAME_ASSETS_DIR, GameProject } from "./gameProject"
 import { GameSound } from "./gameSound"
-import { buildKeyPressEvent, canvasResolution, defaultImgs, queueEvent, resizeCanvas, step, wKeyCode } from "./render"
+import { resizeCanvas, step } from "./render"
+import { buildKeyPressEvent, canvasResolution, defaultImgs, queueEvent, wKeyCode } from "./utils"
 
 export default (project: GameProject, interpreter: Interpreter, canvasParent?: Element) => (p: p5) => {
     const images = new Map<string, p5.Image>()
